# docs/3-reference/rest-openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/eudiw/openid4vp/wallet/simulate" : {
      "post" : {
        "tags" : [ "oid-4vp-controller" ],
        "operationId" : "simulateWallet",
        "parameters" : [ {
          "name" : "verbose",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WalletSimulationPayload"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WalletSimulationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eudiw/openid4vp/validate" : {
      "post" : {
        "tags" : [ "oid-4vp-controller" ],
        "operationId" : "validate",
        "parameters" : [ {
          "name" : "verbose",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ValidationPayload"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eudiw/openid4vp/requests" : {
      "post" : {
        "tags" : [ "oid-4vp-controller" ],
        "operationId" : "createRequest",
        "parameters" : [ {
          "name" : "verbose",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AuthorizationRequestPayload"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthorizationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eudiw/openid4vp/presentations/seed" : {
      "post" : {
        "tags" : [ "oid-4vp-controller" ],
        "operationId" : "seedPresentations",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SeedRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SeedResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/webauthn/attestations/{credentialId}" : {
      "get" : {
        "tags" : [ "web-authn-attestation-stored-controller" ],
        "operationId" : "metadata",
        "parameters" : [ {
          "name" : "credentialId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoredAttestationMetadataResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "VerboseTraceAttributePayload" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "description" : "Attribute type (string, int, hex, etc.)."
          },
          "name" : {
            "type" : "string",
            "description" : "Attribute name."
          },
          "value" : {
            "type" : "string",
            "description" : "Formatted attribute value."
          }
        },
        "description" : "Ordered attribute entry describing a captured value for a verbose trace step."
      },
      "VerboseTracePayload" : {
        "type" : "object",
        "properties" : {
          "operation" : {
            "type" : "string",
            "description" : "Stable identifier describing the traced operation (for example, totp.evaluate.stored)."
          },
          "metadata" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string",
              "description" : "Additional metadata associated with the trace."
            },
            "description" : "Additional metadata associated with the trace."
          },
          "steps" : {
            "type" : "array",
            "description" : "Ordered steps executed during the operation.",
            "items" : {
              "$ref" : "#/components/schemas/VerboseTraceStepPayload"
            }
          }
        },
        "description" : "Deterministic, ordered diagnostic trace emitted when verbose mode is enabled for a credential workflow."
      },
      "VerboseTraceStepPayload" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Stable identifier that enables downstream tooling to group similar steps."
          },
          "summary" : {
            "type" : "string",
            "description" : "Human-readable summary of what the step accomplished."
          },
          "detail" : {
            "type" : "string",
            "description" : "Additional detail about the specific helper or component invoked."
          },
          "spec" : {
            "type" : "string",
            "description" : "Reference to the governing specification section when applicable."
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Arbitrary attributes captured for the step, grouped by attribute type."
            },
            "description" : "Arbitrary attributes captured for the step, grouped by attribute type."
          },
          "orderedAttributes" : {
            "type" : "array",
            "description" : "Ordered attribute entries preserved in capture order.",
            "items" : {
              "$ref" : "#/components/schemas/VerboseTraceAttributePayload"
            }
          },
          "notes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string",
              "description" : "Free-form notes associated with the step."
            },
            "description" : "Free-form notes associated with the step."
          }
        },
        "description" : "Single diagnostic entry representing one action performed while handling the request."
      },
...
      "SeedResponse" : {
        "type" : "object",
        "properties" : {
          "addedCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "canonicalCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "addedCredentialIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "InlineSdJwtPayload" : {
        "type" : "object",
        "properties" : {
          "credentialId" : {
            "type" : "string"
          },
          "format" : {
            "type" : "string"
          },
          "compactSdJwt" : {
            "type" : "string"
          },
          "disclosures" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keyBindingJwt" : {
            "type" : "string"
          },
          "trustedAuthorityPolicies" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "WalletSimulationPayload" : {
        "type" : "object",
        "properties" : {
          "requestId" : {
            "type" : "string"
          },
          "walletPreset" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "responseMode" : {
            "type" : "string"
          },
          "inlineSdJwt" : {
            "$ref" : "#/components/schemas/InlineSdJwtPayload"
          },
          "trustedAuthorityPolicy" : {
            "type" : "string"
          }
        }
      },
      "WalletSimulationResponse" : {
        "required" : [ "presentations", "profile", "requestId", "responseMode", "status", "telemetry" ],
        "type" : "object",
        "properties" : {
          "requestId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "responseMode" : {
            "type" : "string"
          },
          "presentations" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          },
          "trace" : {
            "$ref" : "#/components/schemas/VerboseTracePayload"
          },
          "telemetry" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        }
      },
      "InlineVpTokenPayload" : {
        "type" : "object",
        "properties" : {
          "credentialId" : {
            "type" : "string"
          },
          "format" : {
            "type" : "string"
          },
          "vpToken" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "keyBindingJwt" : {
            "type" : "string"
          },
          "disclosures" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "trustedAuthorityPolicies" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "ValidationPayload" : {
        "type" : "object",
        "properties" : {
          "requestId" : {
            "type" : "string"
          },
          "presetId" : {
            "type" : "string"
          },
          "inlineVpToken" : {
            "$ref" : "#/components/schemas/InlineVpTokenPayload"
          },
          "trustedAuthorityPolicy" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "responseMode" : {
            "type" : "string"
          },
          "inlineDcqlJson" : {
            "type" : "string"
          }
        }
      },
      "AuthorizationRequestPayload" : {
        "type" : "object",
        "properties" : {
          "profile" : {
            "type" : "string"
          },
          "responseMode" : {
            "type" : "string"
          },
          "dcqlPreset" : {
            "type" : "string"
          },
          "dcqlOverride" : {
            "type" : "string"
          },
          "signedRequest" : {
            "type" : "boolean"
          },
          "includeQrAscii" : {
            "type" : "boolean"
          }
        }
      },
      "AuthorizationResponse" : {
        "required" : [ "authorizationRequest", "profile", "requestId", "requestUri", "telemetry" ],
        "type" : "object",
        "properties" : {
          "requestId" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "requestUri" : {
            "type" : "string"
          },
          "authorizationRequest" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "qr" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "trace" : {
            "$ref" : "#/components/schemas/VerboseTracePayload"
          },
          "telemetry" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        }
      },
      "SeedRequest" : {
        "type" : "object",
        "properties" : {
          "source" : {
            "type" : "string"
          },
          "presentations" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "metadata" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        }
      },
...
      "StoredAttestationMetadataResponse" : {
        "type" : "object",
        "properties" : {
          "storedCredentialId" : {
            "type" : "string"
          },
          "attestationId" : {
            "type" : "string"
          },
          "format" : {
            "type" : "string"
          },
          "relyingPartyId" : {
            "type" : "string"
          },
          "origin" : {
            "type" : "string"
          },
          "signingMode" : {
            "type" : "string"
          },
          "certificateChainPem" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "trustAnchorSummaries" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "challenge" : {
            "type" : "string"
          },
          "attestationObject" : {
            "type" : "string"
          },
          "clientDataJson" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}